+++ lib/devpack/gems.rb
@@ -3,6 +3,8 @@
 module Devpack
   # Loads requested gems from configuration
   class Gems
+    include Timeable
+
     def initialize(config)
       @config = config
     end
@@ -12,17 +14,11 @@ module Devpack
 
       gems, time = timed { load_devpack }
       names = gems.map(&:first)
-      warn(Messages.loaded_message(@config.devpack_path, gems, time.round(2)))
+      warn(Messages.loaded(@config.devpack_path, gems, time.round(2)))
       names
     end
 
     private
-
-    def timed
-      start = Time.now.utc
-      result = yield
-      [result, Time.now.utc - start]
-    end
 
     def load_devpack
       @config.requested_gems.map { |name| load_gem(name) }.compact
@@ -32,12 +28,18 @@ module Devpack
       update_load_path(name)
       [name, Kernel.require(name)]
     rescue LoadError => e
-      warn(Messages.failure_message(name, e))
+      warn(Messages.failure(name, load_error_message(e)))
       nil
     end
 
     def warn(message)
-      Kernel.warn("[devpack] #{message}")
+      Devpack.warn(message)
+    end
+
+    def load_error_message(error)
+      return "(#{error.message})" unless Devpack.debug?
+
+      %[(#{error.message})\n#{error.backtrace.join("\n")}]
     end
 
     def gem_glob