+++ lib/devpack/messages.rb
@@ -3,18 +3,34 @@
 module Devpack
   # Generates output messages.
   class Messages
-    def self.failure_message(name, error)
-      base = "Failed to load `#{name}`"
-      install = "bundle exec gem install #{name}"
-      "#{base}. Try `#{install}` (#{error})"
-    end
+    class << self
+      def failure(name, error_message)
+        base = "Failed to load `#{name}`"
+        install = "bundle exec gem install #{name}"
+        "#{base}. Try `#{install}` #{error_message}"
+      end
 
-    def self.loaded_message(path, gems, time)
-      already_loaded = gems.size - gems.reject { |_, loaded| loaded }.size
-      base = "Loaded #{already_loaded} development gem(s) from '#{path}' in #{time} seconds"
-      return "#{base}." if already_loaded == gems.size
+      def initializer_failure(path, error_message)
+        "Failed to load initializer `#{path}`: #{error_message}"
+      end
 
-      "#{base} (#{gems.size - already_loaded} gem(s) were already loaded by environment)."
+      def loaded(path, gems, time)
+        already_loaded = gems.size - gems.reject { |_, loaded| loaded }.size
+        base = "Loaded #{already_loaded} development gem(s) from '#{path}' in #{time} seconds"
+        return "#{base}." if already_loaded == gems.size
+
+        "#{base} (#{gems.size - already_loaded} gem(s) were already loaded by environment)."
+      end
+
+      def loaded_initializers(path, initializers, time)
+        "Loaded #{initializers.compact.size} initializer(s) from '#{path}' in #{time} seconds"
+      end
+
+      private
+
+      def indented(message)
+        message.split("\n").map { |line| "  #{line}" }.join("\n")
+      end
     end
   end
 end