+++ spec/devpack/gem_path_spec.rb
@@ -3,7 +3,7 @@
 RSpec.describe Devpack::GemPath do
   subject(:gem_path) { described_class.new(glob, name) }
 
-  let(:glob) { double(find: Pathname.new(Dir.tmpdir)) }
+  let(:glob) { instance_double(Devpack::GemGlob, find: Pathname.new(Dir.tmpdir)) }
   let(:name) { 'example' }
 
   it { is_expected.to be_a described_class }