+++ spec/devpack/gems_spec.rb
@@ -1,20 +1,24 @@
 # frozen_string_literal: true
 
-# REVIEW: Checking logs by expecting `Kernel.warn` to be called an exact amount
-# of times will break if any other warnings occur. We should ensure that each
-# message starts with "[devpack]".
 RSpec.describe Devpack::Gems do
   subject(:gems) { described_class.new(config) }
 
-  let(:config) { Devpack::Config.new(project_path) }
+  let(:config) do
+    instance_double(
+      Devpack::Config,
+      requested_gems: requested_gems,
+      devpack_path: devpack_path.join('.devpack')
+    )
+  end
   let(:project_path) { Pathname.new(Dir.tmpdir).join('example') }
   let(:devpack_path) { project_path }
+  let(:requested_gems) { [] }
   let(:gem_home) { Pathname.new(File.join(Dir.tmpdir, 'gem_home')) }
   it { is_expected.to be_a described_class }
   its(:load) { is_expected.to be_an Array }
 
   describe '#load' do
-    subject(:load) { gems.load }
+    subject(:gems_load) { gems.load }
 
     let(:installed_gems) { %w[installed1 installed2 installed3] }
     let(:not_installed_gems) { %w[not_installed1 not_installed2 not_installed3] }
@@ -22,32 +26,48 @@ RSpec.describe Devpack::Gems do
     before do
       stub_const('ENV', ENV.to_h.merge('GEM_PATH' => "#{gem_home}:/some/other/directory"))
       FileUtils.mkdir_p(project_path)
-      File.write(devpack_path.join('.devpack'), requested_gems.join("\n"))
       allow(Kernel).to receive(:require).and_call_original
       installed_gems.each { |name| allow(Kernel).to receive(:require).with(name) }
     end
 
-    after { FileUtils.rm_r(devpack_path) }
-
     context 'with .devpack file in provided directory' do
       context 'with all specified gems installed' do
-        before { expect(Kernel).to receive(:warn).exactly(1).times }
+        before { expect(Devpack).to receive(:warn).exactly(1).times }
         let(:requested_gems) { installed_gems }
         it { is_expected.to eql requested_gems }
       end
 
       context 'with some specified gems not installed' do
         let(:requested_gems) { installed_gems + not_installed_gems }
-        before { expect(Kernel).to receive(:warn).exactly(4).times }
+        before { expect(Devpack).to receive(:warn).exactly(4).times }
         it { is_expected.to eql installed_gems }
+      end
+
+      context 'with missing gems and DEVPACK_DEBUG enabled' do
+        let(:requested_gems) { %w[not_installed1] }
+        before { allow(Devpack).to receive(:debug?) { true } }
+        it 'issues a warning including error and traceback' do
+          expect(Devpack).to receive(:warn).at_least(:once).with(any_args) do |message|
+            next if message.start_with?('Loaded 0 development gem(s)')
+
+            [
+              "/devpack/lib/devpack/gems.rb:29:in `load_gem'",
+              "/devpack/lib/devpack/gems.rb:24:in `block in load_devpack'",
+              "/devpack/lib/devpack/gems.rb:24:in `map'",
+              "/devpack/lib/devpack/gems.rb:24:in `load_devpack'",
+              "/devpack/lib/devpack/gems.rb:15:in `block in load'",
+              "/devpack/lib/devpack/gems.rb:15:in `load'"
+            ].each { |line| expect(message).to include line }
+          end
+          subject
+        end
       end
     end
 
     context 'no .devpack file present in provided directory' do
-      let(:requested_gems) { [] }
+      let(:requested_gems) { nil }
       before do
-        expect(Kernel).to_not receive(:warn)
-        File.delete(devpack_path.join('.devpack'))
+        expect(Devpack).to_not receive(:warn)
       end
       it { is_expected.to be_empty }
     end
@@ -70,7 +90,7 @@ RSpec.describe Devpack::Gems do
 
       after { FileUtils.rm_r(gem_home.to_s) }
 
-      before { expect(Kernel).to receive(:warn).exactly(1).times }
+      before { expect(Devpack).to receive(:warn).exactly(1).times }
       it { is_expected.to eql requested_gems }
     end
   end